Move benchmark() behind a BenchmarkTools/PrettyTables extension
BenchmarkTools and PrettyTables were hard dependencies used only by the
benchmark() utility, taxing load time for every detection-only user.
They are now weakdeps triggering a BenchmarkExt extension; calling
benchmark() without them loaded raises a MethodError with a hint
explaining what to load.
