+++ Project.toml
@@ -9,7 +9,6 @@ projects = ["test"]
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 AllocArrays = "5c00bae2-1499-4716-9206-27f63fd08a44"
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
@@ -19,15 +18,17 @@ ImageFiltering = "6a3955dd-da59-5b1f-98d4-e7296123deb5"
 ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
-PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 UnsafeArrays = "c4a57d5a-5b31-53a6-b365-19f8c011fbd6"
 
 [weakdeps]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 cuDNN = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
+PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 
 [extensions]
+BenchmarkExt = ["BenchmarkTools", "PrettyTables"]
 CUDAExt = ["CUDA", "cuDNN"]
 
 [compat]
